+++ packages/gov-ui-kit/src/components/input/index.ts
@@ -1,5 +1,6 @@
 export * from './inputContainer';
 export * from './inputDate';
+export * from './inputNumber';
 export * from './inputNumberMax';
 export * from './inputSearch';
 export * from './inputText';