+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,10 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+-   Implement `Link`, `InputNumber` and `InputTime` components
 -   Implement 'Addon' element for `InputText` component
--   Implement `Link` component
 -   Handle size property on `Progress` component
--   Implement `InputTime` component
 
 ### Changed